zlib.h: Use unsigned long for totals, for compatibility with zlib on 64-bit Windows
Stock zlib uses unsigned long for total_in and total_out. On 64-bit
Windows, that isn't the same as size_t.

+++ zlib.h
@@ -94,11 +94,11 @@ struct internal_state;
 typedef struct z_stream_s {
     const unsigned char   *next_in;   /* next input byte */
     uint32_t              avail_in;   /* number of bytes available at next_in */
-    size_t                total_in;   /* total number of input bytes read so far */
+    unsigned long         total_in;   /* total number of input bytes read so far */
 
     unsigned char         *next_out;  /* next output byte will go here */
     uint32_t              avail_out;  /* remaining free space at next_out */
-    size_t                total_out;  /* total number of bytes output so far */
+    unsigned long         total_out;  /* total number of bytes output so far */
 
     const char            *msg;       /* last error message, NULL if no error */
     struct internal_state *state;     /* not visible by applications */
